:memo: Add documentation to explain how to correctly Browserify the okta-signin-widget module
When developers tried to browserify an app that require the okta-signin-widget, the bundling failed because Browserify attempts to resolve paths in require() calls in the already bundled okta-signin-widget, while it should ignore them since the okta-sign-in.entry.js is a standalone file and doesn't need the external modules, they are already bundled in the same file. Added documentation to explain how to tell to browserify to not parse the okta-signin-widget module and run correctly.

Resolves: OKTA-135028