:bug: Add missing table row element
In Mobile iOs devices, pressing the middle option for enrolling devices in Okta Verify was freezing the browser and the Okta Mobile app.
There is a missing table-row element, which in theory browsers will add automatically (reason it works everywhere else) but seems that iOs is not doing that. So to prevent any future error, I am explicitly adding the row.

Resolves: OKTA-155395