:bug: Fixes Okta Verify polling issue when switching factors (#404)
* We continute to poll even after user navigates to another factor after
triggering a MFA verify. This was due to the fact that the transaction state not being set
when sing MFA verify until the poll complted with a success/failure.

The fix here is that the transaction is updated when we get the first response from server.

* used manageTransaction instead of doTransaction to allow to set the transaction and then
continue polling in okta verify case.

* Fixed an issue in manageTranaction. We  now return the promise else we wrap the result in
a promise.