fix: Handle input[type="tel"] the same way as other text input types. (#625)
This is a CSS fix to complete PR #367. That PR introduced the use of type="tel" input fields. This PR updates the CSS to treat those fields in the same way as e.g. "text", "number" and "password" input fields -- in particular, the entire input field (including the right-hand part) should be clickable, not just the left-hand part.

Part of:  OKTA-201572