fix: Prevents redirect when app is in background (#787)
Introduces a flag to check if SIW host is in background before redirecting.

This is to overcome the issue where Chrome on android prevents app url redirects when not in foreground. This is currently happening in chrome but not on firefox on android. IOS is not an issue as safari takes care of it.

Resolves: OKTA-235748